requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==1.12.0
python-dotenv==1.0.0
pandas==2.1.3
requests==2.31.0
Werkzeug==2.3.7
httpx==0.25.0
tiktoken==0.5.2